[UPDATE] DTO

+++ src/main/java/com/treemarket/tree/mapper/CategoryMapper.java
@@ -1,11 +1,8 @@
 package com.treemarket.tree.mapper;
 
-import com.treemarket.tree.domain.CategoryVO;
 import org.apache.ibatis.annotations.Mapper;
 
 @Mapper
 public interface CategoryMapper {
-
-    long getCtgId(String ctgName);
-
+    int getCtgId (String ctgName);
 }
