+++ src/main/java/com/treemarket/tree/service/CategoryService.java
@@ -1,9 +1,5 @@
 package com.treemarket.tree.service;
 
-import com.treemarket.tree.domain.CategoryVO;
-
 public interface CategoryService {
-
-    long getCtgId(String ctgName);
-
+    int getCtgId (String ctgName);
 }
